hyper_util::client::legacy

Module client

The legacy HTTP Client from 0.14.x

This Client will eventually be deconstructed into more composable parts. For now, to enable people to use hyper 1.0 quicker, this Client exists in much the same way it did in hyper 0.14.

Macros

e 🔒

Structs

Builder

A builder to configure a new Client.

Client

A Client to make outgoing HTTP requests.

Config 🔒

Error

Client errors

PoolClient 🔒

ResponseFuture

A Future that will resolve to an HTTP Response.

Enums

ClientConnectError 🔒

ErrorKind 🔒

PoolTx 🔒

TrySendError 🔒

Functions

absolute_form 🔒

authority_form 🔒

domain_as_uri 🔒

extract_domain 🔒

get_non_default_port 🔒

is_schema_secure 🔒

origin_form 🔒

set_scheme 🔒

Type Aliases

BoxSendFuture 🔒

PoolKey 🔒