hyper_util::client::proxy

Module matcher

Proxy matchers

This module contains different matchers to configure rules for when a proxy should be used, and if so, with what arguments.

A Matcher can be constructed either using environment variables, or a Matcher::builder().

Once constructed, the Matcher can be asked if it intercepts a Uri by calling Matcher::intercept().

An Intercept includes the destination for the proxy, and any parsed authentication to be used.

Structs

Builder

A builder to create a Matcher.

Intercept

A matched proxy,

Matcher

A proxy matcher, usually built from environment variables.