hyper_util::client::proxy::matcher

Struct Matcher

pub struct Matcher { /* private fields */ }

A proxy matcher, usually built from environment variables.

Implementations

impl Matcher

pub fn from_env() -> Self

Create a matcher reading the current environment variables.

This checks for values in the following variables, treating them the same as curl does:

ALL_PROXY/all_proxy

HTTPS_PROXY/https_proxy

HTTP_PROXY/http_proxy

NO_PROXY/no_proxy

pub fn from_system() -> Self

Create a matcher from the environment or system.

This checks the same environment variables as from_env(), and if not set, checks the system configuration for values for the OS.

This constructor is always available, but if the client-proxy-system feature is enabled, it will check more configuration. Use this constructor if you want to allow users to optionally enable more, or use from_env if you do not want the values to change based on an enabled feature.

pub fn builder() -> Builder

Start a builder to configure a matcher.

pub fn intercept(&self, dst: &Uri) -> Option<Intercept>

Check if the destination should be intercepted by a proxy.

If the proxy rules match the destination, a new Uri will be returned to connect to.