hyper_util::client::proxy::matcher

Trait IntoValue

pub trait IntoValue { }

Available on crate feature client and crate feature client-proxy and docsrs only.

A type that can used as a Builder value.

Private and sealed, only visible in docs.