Struct hyper_util::server::conn::auto::Http1Builder

pub struct Http1Builder<'a, E> { /* private fields */ }

Http1 part of builder.

Implementations

impl<E> Http1Builder<'_, E>

pub fn http2(&mut self) -> Http2Builder<'_, E>

Http2 configuration.

pub fn half_close(&mut self, val: bool) -> &mut Self

Set whether HTTP/1 connections should support half-closures.

Clients can chose to shutdown their write-side while waiting for the server to respond. Setting this to true will prevent closing the connection immediately if read detects an EOF in the middle of a request.

Default is false.

pub fn keep_alive(&mut self, val: bool) -> &mut Self

Enables or disables HTTP/1 keep-alive.

Default is true.

pub fn title_case_headers(&mut self, enabled: bool) -> &mut Self

Set whether HTTP/1 connections will write header names as title case at the socket level.

Note that this setting does not affect HTTP/2.

Default is false.

pub fn preserve_header_case(&mut self, enabled: bool) -> &mut Self

Set whether to support preserving original header cases.

Currently, this will record the original cases received, and store them in a private extension on the Request. It will also look for and use such an extension in any provided Response.

Since the relevant extension is still private, there is no way to interact with the original cases. The only effect this can have now is to forward the cases in a proxy-like fashion.

Note that this setting does not affect HTTP/2.

Default is false.

pub fn max_headers(&mut self, val: usize) -> &mut Self

Set the maximum number of headers.

When a request is received, the parser will reserve a buffer to store headers for optimal performance.

If server receives more headers than the buffer size, it responds to the client with “431 Request Header Fields Too Large”.

The headers is allocated on the stack by default, which has higher performance. After setting this value, headers will be allocated in heap memory, that is, heap memory allocation will occur for each request, and there will be a performance drop of about 5%.

Note that this setting does not affect HTTP/2.

Default is 100.

pub fn header_read_timeout( &mut self, read_timeout: impl Into<Option<Duration>>, ) -> &mut Self

Set a timeout for reading client request headers. If a client does not transmit the entire header within this time, the connection is closed.

Requires a Timer set by Http1Builder::timer to take effect. Panics if header_read_timeout is configured without a Timer.

Pass None to disable.

Default is currently 30 seconds, but do not depend on that.

pub fn writev(&mut self, val: bool) -> &mut Self

Set whether HTTP/1 connections should try to use vectored writes, or always flatten into a single buffer.

Note that setting this to false may mean more copies of body data, but may also improve performance when an IO transport doesn’t support vectored writes well, such as most TLS implementations.

Setting this to true will force hyper to use queued strategy which may eliminate unnecessary cloning on some TLS backends

Default is auto. In this mode hyper will try to guess which mode to use

pub fn max_buf_size(&mut self, max: usize) -> &mut Self

Set the maximum buffer size for the connection.

Default is ~400kb.

Panics

The minimum value allowed is 8192. This method panics if the passed max is less than the minimum.

pub fn pipeline_flush(&mut self, enabled: bool) -> &mut Self

Aggregates flushes to better support pipelined responses.

Experimental, may have bugs.

Default is false.

pub fn timer<M>(&mut self, timer: M) -> &mut Self
where M: Timer + Send + Sync + 'static,

Set the timer used in background tasks.

pub async fn serve_connection<I, S, B>( &self, io: I, service: S, ) -> Result<(), Box<dyn Error + Send + Sync>>
where S: Service<Request<Incoming>, Response = Response<B>>, S::Future: 'static, S::Error: Into<Box<dyn StdError + Send + Sync>>, B: Body + 'static, B::Error: Into<Box<dyn StdError + Send + Sync>>, I: Read + Write + Unpin + 'static, E: HttpServerConnExec<S::Future, B>,

Bind a connection together with a Service.

pub fn serve_connection_with_upgrades<I, S, B>( &self, io: I, service: S, ) -> UpgradeableConnection<'_, I, S, E>
where S: Service<Request<Incoming>, Response = Response<B>>, S::Future: 'static, S::Error: Into<Box<dyn StdError + Send + Sync>>, B: Body + 'static, B::Error: Into<Box<dyn StdError + Send + Sync>>, I: Read + Write + Unpin + Send + 'static, E: HttpServerConnExec<S::Future, B>,

Bind a connection together with a Service, with the ability to handle HTTP upgrades. This requires that the IO object implements Send.