hyper_util::server

Module graceful

Available on crate features server and server-graceful only.

Utility to gracefully shutdown a server.

This module provides a GracefulShutdown type, which can be used to gracefully shutdown a server.

See https://github.com/hyperium/hyper-util/blob/master/examples/server_graceful.rs for an example of how to use this.

Structs

GracefulShutdown

A graceful shutdown utility

Watcher

A watcher side of the graceful shutdown.

Traits

GracefulConnection

An internal utility trait as an umbrella target for all (hyper) connection types that the GracefulShutdown can watch.