hyper_util::server::graceful

Trait GracefulConnection

pub trait GracefulConnection: Future<Output = Result<(), Self::Error>> + Sealed {
    type Error;

    // Required method
    fn graceful_shutdown(self: Pin<&mut Self>);
}

An internal utility trait as an umbrella target for all (hyper) connection types that the GracefulShutdown can watch.

Required Associated Types

type Error

The error type returned by the connection when used as a future.

Required Methods

fn graceful_shutdown(self: Pin<&mut Self>)

Start a graceful shutdown process for this connection.