hyper_util::client::legacy

Module client

The legacy HTTP Client from 0.14.x

This Client will eventually be deconstructed into more composable parts. For now, to enable people to use hyper 1.0 quicker, this Client exists in much the same way it did in hyper 0.14.

Macros

e ğŸ”’

Structs

Builder

A builder to configure a new Client.

Client

A Client to make outgoing HTTP requests.

Config ğŸ”’

Error

Client errors

PoolClient ğŸ”’

ResponseFuture

A Future that will resolve to an HTTP Response.

Enums

ClientConnectError ğŸ”’

ErrorKind ğŸ”’

PoolTx ğŸ”’

TrySendError ğŸ”’

Functions

absolute_form ğŸ”’

authority_form ğŸ”’

domain_as_uri ğŸ”’

extract_domain ğŸ”’

get_non_default_port ğŸ”’

is_schema_secure ğŸ”’

origin_form ğŸ”’

set_scheme ğŸ”’

Type Aliases

BoxSendFuture ğŸ”’

PoolKey ğŸ”’