hyper_util::client::proxy::matcher

Struct Builder

pub struct Builder { /* private fields */ }

Available on crate features client and client-proxy only.

A builder to create a Matcher.

Construct with Matcher::builder().

Implementations

impl Builder

pub fn all<S>(self, val: S) -> Self
where S: IntoValue,

Set the target proxy for all destinations.

pub fn http<S>(self, val: S) -> Self
where S: IntoValue,

Set the target proxy for HTTP destinations.

pub fn https<S>(self, val: S) -> Self
where S: IntoValue,

Set the target proxy for HTTPS destinations.

pub fn no<S>(self, val: S) -> Self
where S: IntoValue,

Set the “no” proxy filter.

The rules are as follows:

Entries are expected to be comma-separated (whitespace between entries is ignored)

IP addresses (both IPv4 and IPv6) are allowed, as are optional subnet masks (by adding /size, for example “192.168.1.0/24”).

An entry “*” matches all hostnames (this is the only wildcard allowed)

Any other entry is considered a domain name (and may contain a leading dot, for example google.com and .google.com are equivalent) and would match both that domain AND all subdomains.

For example, if "NO_PROXY=google.com, 192.168.1.0/24" was set, all of the following would match (and therefore would bypass the proxy):

http://google.com/

http://www.google.com/

http://192.168.1.42/

The URL http://notgoogle.com/ would not match.

pub fn build(self) -> Matcher

Construct a Matcher using the configured values.